An ADO.NET contract fixture says which database it is, and nothing else
The SQLite contract fixture carried the whole recipe for running the job
store contract against a real database: build the provider, build the
store, initialize it, and never call SchedulerStarted. Five more dialects
want the same recipe, and copying it five times would let the copies
drift apart on exactly the discipline that makes the run meaningful.

AdoJobStoreContractTest holds it once. A subclass supplies a provider
name, a connection string and a driver delegate; the base derives a
scheduler name from the subclass's own type so no two fixtures can see
each other's rows, clears the store as it builds it so a container's
leftovers cannot be counted by the next test, and passes no lock handler
so the store picks the one its dialect needs — which is the choice
production leaves it to make.

The connection strings the container fixtures need move to TestConstants
beside the two that already lived there, so the paging test stops
carrying its own copies of them.
